Run Complete DESeq2 Pipeline

This function runs a complete RNA-seq analysis pipeline including quality control, differential expression analysis, and gene set enrichment analysis for all possible comparisons in the experiment. It generates various plots and saves results in organized directories.

Usage

run_deseq2_pip(
  dds,
  experiment,
  batch = NULL,
  save_dir = getwd(),
  org = "human",
  remove_xy = FALSE,
  remove_mt = FALSE,
  group_by = "Group1",
  quantile = 0.05,
  order = "rank",
  save_dir_name = "qc_results",
  assaytype = "RNA"
)

Arguments

dds

DESeq2 object containing the expression data

experiment

Name of the experiment

batch

Column name in colData(dds) to use for batch correction. Default is NULL

save_dir

Directory where all output files will be saved. Default is current working directory

org

The organism to use, either "human" or "mouse". Default is "human"

remove_xy

Logical. If TRUE, removes genes on X and Y chromosomes. Default is FALSE

remove_mt

Logical. If TRUE, removes mitochondrial genes. Default is FALSE

group_by

Column name in colData(dds) to use for grouping. Default is "Group1"

quantile

Quantile threshold for filtering lowly expressed genes. Default is 0.05

order

Column name to use for ranking genes. Default is "rank"

save_dir_name

Name of the subdirectory to save files in. Default is "qc_results"

assaytype

Type of assay, either "RNA" or "ATAC". Default is "RNA"

Value

The processed DESeq2 object

Examples

if (FALSE) { # \dontrun{
# Run complete pipeline with default settings
dds <- run_deseq2_pip(dds, experiment = "my_experiment")

# Run pipeline with custom settings
dds <- run_deseq2_pip(dds, experiment = "my_experiment",
                    org = "mouse", remove_xy = TRUE,
                    group_by = "Treatment",
                    save_dir_name = "custom_results")

# Run pipeline with batch correction
dds <- run_deseq2_pip(dds, experiment = "my_experiment",
                    batch = "Batch")
} # }